Scripts/requirements.txt
aiohttp==3.6.2
apply-defaults==0.1.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==6.7
cycler==0.10.0
dict2xml==1.5
discord.py==1.3.1
idna==2.8
js2xml==0.3.1
json2xml==3.0.1
jsonrpcclient==3.3.5
jsonschema==3.2.0
kiwisolver==1.1.0
lxml==4.4.2
matplotlib==3.1.3
multidict==4.7.4
nltk==3.4.5
numpy==1.17.4
pandas==0.25.3
Pillow==7.0.0
ply==3.11
pyaes==1.6.1
pyasn1==0.4.8
pycoingecko==1.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
slimit==0.8.1
soupsieve==1.9.5
Telethon==1.10.10
urllib3==1.25.7
websockets==8.1
wordcloud==1.6.0
xmltodict==0.11.0
yarl==1.4.2
